+++ grade.h
@@ -1,0 +1,11 @@
+#ifndef GUARD_grade_h
+#define GUARD_grade_h
+#include <vector>
+#include <string>
+#include "Student_info.h"
+
+
+double grade (double midterm, double final, double homework);
+double grade (double midterm, double final, const std::vector<double>& hw);
+double grade (const Student_info& s);
+#endif
