+++ median.h
@@ -1,0 +1,8 @@
+
+#ifndef GUARD_median_h
+#define GUARD_median_h
+#include <vector>
+
+double median (std::vector<double> vec);
+
+#endif
